разделила на компоненты favorites-page
