+++ src/pages/favorites-page/favorites-page.tsx
@@ -1,57 +1,11 @@
-import CitiesInfo from '../../components/general/header';
+import Header from '../../components/general/header';
 import Nav from '../../components/general/nav';
 import {Offer} from '../../index/index';
-import PlaceCard from '../../components/ui/offer-card';
-import {CITIES} from '../../const';
 import {Helmet} from 'react-helmet-async';
-
-// правильный ли подход используется к компонентам в одном файле, снизу-вверх?
+import FavoritesList from '../../components/favorite/favorites-list';
 
 type FavoritesProps = {
   offers: Offer[];
-}
-
-function FavoritesPlaces({offers}: FavoritesProps): (JSX.Element | null)[] | null {
-  return (
-    offers.map((item: Offer): JSX.Element|null => {
-      const {images, isFavorite} = item;
-      return isFavorite ?
-        <article key={item.id} className="favorites__card place-card">
-          <div className="favorites__image-wrapper cities__image-wrapper place-card__image-wrapper">
-            <a href="#">
-              {images && <img className="place-card__image" src={images[0]} width={260} height={200} alt="Place image"/>}
-            </a>
-          </div>
-          <PlaceCard offer={item}/>
-        </article>
-        : null;
-    })
-  );
-}
-
-// eslint иначе не пропускал(() => {})()
-// сделала пока так, наверное потом переделаю после ретро
-function FavoritesList({offers}: FavoritesProps): JSX.Element | null {
-  const favoriteslocations = (!offers.length) ? null : (
-    <ul className="favorites__list">
-      {CITIES.map((item) => (() => {
-        const filterCards = offers.filter((el) => el.city.name === item && el.isFavorite);
-        return (filterCards.length) ?
-          <li key={item} className="favorites__locations-items">
-            <div className="favorites__locations locations locations--current">
-              <div className="locations__item">
-                <a className="locations__item-link" href="#">
-                  <span>{item}</span>
-                </a>
-              </div>
-            </div>
-            <div className="favorites__places">
-              <FavoritesPlaces offers={filterCards} />
-            </div>
-          </li> : null;
-      })())}
-    </ul>);
-  return favoriteslocations;
 }
 
 function FavoritePage({offers} :FavoritesProps): JSX.Element {
@@ -60,9 +14,9 @@ function FavoritePage({offers} :FavoritesProps): JSX.Element {
       <Helmet>
         <title>{'6 cities - Favorite'}</title>
       </Helmet>
-      <CitiesInfo>
+      <Header>
         <Nav/>
-      </CitiesInfo>
+      </Header>
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
           <section className="favorites">
